backend/database/mongo_config.py
import os
from motor.motor_asyncio import AsyncIOMotorClient
from dotenv import load_dotenv
from pathlib import Path

# Load .env from backend directory
env_path = Path(__file__).resolve().parent.parent / '.env'
load_dotenv(dotenv_path=env_path)

MONGODB_URL = os.getenv("MONGODB_URL")

class MongoDB:
    client: AsyncIOMotorClient = None
    db = None

db = MongoDB()

async def connect_to_mongo():
    # Adding tlsAllowInvalidCertificates=True to bypass SSL errors in dev environment
    db.client = AsyncIOMotorClient(MONGODB_URL, tlsAllowInvalidCertificates=True)
    db.db = db.client.get_database("shield_db")
    print("Connected to MongoDB")

async def close_mongo_connection():
    db.client.close()
    print("Closed MongoDB connection")

def get_database():
    return db.db
